pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent_foundry"
version = "0.1.0"
authors = [
    { name="Justin O'Connor", email="justin@example.com" },
]
description = "A framework for building and managing AI agents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "click",
    "semantic-kernel",
    "python-dotenv",
    "aiohttp",
    "jsonschema"
]

[project.urls]
Homepage = "https://github.com/justinoconnor/agent-foundry"
Repository = "https://github.com/justinoconnor/agent-foundry"

[project.scripts]
odk = "agent_foundry.cli.cli:cli"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
packages = ["agent_foundry"]

[[tool.mypy.overrides]]
module = "semantic_kernel.*"
ignore_missing_imports = true

[tool.rye]
managed = true
dev-dependencies = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio",
    "click",
    "black",
    "flake8",
    "flake8-docstrings",
    "isort",
    "mypy",
    "pre-commit",
    "pydantic",
    "types-requests",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["agent_runtime"]
